rename `.net` to `.network`
(leaving `rest.net` as this pertains to the Net API)
